todo : fix mutable error

+++ src/app/dashboard/home/home.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+import { Component, inject, OnInit, ViewEncapsulation } from '@angular/core';
 import { RouterLink, RouterOutlet } from '@angular/router';
 import { Observable } from 'rxjs';
 import { CollectionRequest } from '../../model/collection-request.model';
@@ -7,6 +7,7 @@ import { selectAllCollections } from '../../store/collection.selectors';
 import { CommonModule } from '@angular/common';
 import { CollectionActions } from '../../store/collection.actions';
 import { FormsModule } from '@angular/forms';
+import { AuthService } from '../../services/auth.service';
 
 @Component({
   selector: 'app-home',
@@ -29,7 +30,7 @@ export class HomeComponent implements OnInit {
   
   updateStatusModalVisible: boolean = false;
   collectionToUpdate: CollectionRequest | null = null;
-
+  authService = inject(AuthService)
 
 
 
@@ -39,13 +40,21 @@ export class HomeComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    const userString = localStorage.getItem('user');
+    const userString = localStorage.getItem('current_user');
     if (userString) {
       const user = JSON.parse(userString);
       this.role = user.role;
     }
   }
+  // editCollection(collection: CollectionRequest): void {
+  //   this.collectionToEdit = { ...collection };
+  //   this.editModalVisible = true;
+  // }
   editCollection(collection: CollectionRequest): void {
+    if (this.role === 'collector' && collection.status !== 'en attente') {
+      alert('Collectors can only edit pending requests.');
+      return;
+    }
     this.collectionToEdit = { ...collection };
     this.editModalVisible = true;
   }
@@ -62,6 +71,14 @@ export class HomeComponent implements OnInit {
   }
 
   updateStatus(collection: CollectionRequest): void {
+    if (this.role !== 'collector') {
+      alert('Only collectors can update the status.');
+      return;
+    }
+    if (collection.status === 'validee') {
+      alert('This request is validated and cannot be updated.');
+      return;
+    }
     this.collectionToUpdate = { ...collection };
     this.updateStatusModalVisible = true;
   }
@@ -76,11 +93,23 @@ export class HomeComponent implements OnInit {
       this.closeStatusModal();
     }
   }
-  removeCollection(collectionId: string): void {
-    if (confirm("Are you sure you want to remove this collection?")) {
+  // removeCollection(collectionId: string): void {
+  //   if (confirm("Are you sure you want to remove this collection?")) {
+  //     this.store.dispatch(CollectionActions.removeCollection({ collectionid: collectionId }));
+  //   }
+  // }
+  removeCollection(collectionId: string, collectionStatus: string): void {
+    if (this.role === 'collector' && collectionStatus !== 'en attente') {
+      alert('Collectors can only remove pending requests.');
+      return;
+    }
+    if (confirm('Are you sure you want to remove this collection?')) {
       this.store.dispatch(CollectionActions.removeCollection({ collectionid: collectionId }));
     }
   }
-
+  logout(): void {
+    this.authService.logout();
+    console.log('logout');
+  }
 
 }
